projects/TropixStorageCore/src/fs-impl/edu/umn/msi/tropix/storage/core/access/fs/CascadingFunctionImpl.java
/*******************************************************************************
 * Copyright 2009 Regents of the University of Minnesota. All rights
 * reserved.
 * Copyright 2009 Mayo Foundation for Medical Education and Research.
 * All rights reserved.
 *
 * This program is made available under the terms of the Eclipse
 * Public License v1.0 which accompanies this distribution,
 * and is available at http://www.eclipse.org/legal/epl-v10.html
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, EITHER EXPRESS OR
 * IMPLIED INCLUDING, WITHOUT LIMITATION, ANY WARRANTIES OR CONDITIONS
 * OF TITLE, NON-INFRINGEMENT, MERCHANTABILITY OR FITNESS FOR A
 * PARTICULAR PURPOSE.  See the License for the specific language
 * governing permissions and limitations under the License.
 *
 * Contributors:
 * Minnesota Supercomputing Institute - initial API and implementation
 ******************************************************************************/

package edu.umn.msi.tropix.storage.core.access.fs;

import com.google.common.base.Function;

class CascadingFunctionImpl<S, T> implements Function<S, T> {
  private Iterable<? extends Function<S, T>> functions; 
  
  CascadingFunctionImpl(Iterable<? extends Function<S, T>> functions) {
    this.functions = functions;
  }
  
  public T apply(S input) {
    T result = null;
    for(Function<S, T> function : functions) {
      result = function.apply(input);
      if(result != null) {
        break;
      }
    }
    return result;
  }

}
